Tessera had a frame and nothing that said what its numbers meant. A view
now declares a projection from a closed set — web_mercator,
equirectangular and its aliases, or none — states its extent in longitude
and latitude, and the build and the write path both transform at the
boundary through that one function.

Eight phases. The coordinate path is f64 end to end, which removed a
narrowing that had been acking coordinates outside the declared extent.
The transform normalises to the unit square with y south, held to the
Python module that placed the corpora. The frame is the projection's
domain or a 2^k-aligned sub-square, snapped outward from a box written in
degrees. The report names the projection and counts clipped points
separately from clamped, the two having separate causes. /v1/meta
publishes the tile scheme a frame addresses, which is what decides whether
a basemap can be drawn — an equirectangular frame is aligned to a tiling
no server serves. Shapes may be declared in longitude and latitude, their
edges straight in the plane they were declared in and densified before
projection. Both geographic corpora are rebuilt: every one of 87,094,949
stored positions agrees with an independent recomputation.

Amends configuration.md §1 and contracts.md §2.2, §3.2 and §3.4;
polygon-membership.md reaches r8 with its wgs84 blockage removed.

+++ crates/tessera-cli/tests/corpus_cli.rs
@@ -6,7 +6,7 @@
 use std::io::Write;
 use std::process::{Command, Stdio};
 
-use arrow::array::{Float32Array, UInt64Array};
+use arrow::array::{Float64Array, UInt64Array};
 use arrow::record_batch::RecordBatch;
 use tessera_corpus::{Corpus, Grant};
 use tessera_spatial::Bounds;
@@ -74,7 +74,7 @@ fn items_answers_served_keys_with_their_items() {
         .column_by_name("x")
         .unwrap()
         .as_any()
-        .downcast_ref::<Float32Array>()
+        .downcast_ref::<Float64Array>()
         .unwrap();
     for (i, key) in keys.iter().enumerate() {
         assert_eq!(fx.value(i), *key);
