+++ crates/tessera-engine/tests/merge.rs
@@ -206,7 +206,7 @@ fn flush_interleaved_segments(engine: &Engine) -> Vec<Vec<(EntityId, String)>> {
                 external_id: Some(external_id.as_bytes().to_vec()),
                 view: "s0".to_string(),
                 // x ≡ s (mod TIER_WIDTH), scaled to distinct cells inside the extent.
-                x: ((t * TIER_WIDTH + s) * 20) as f32,
+                x: ((t * TIER_WIDTH + s) * 20) as f64,
                 y: 5.0,
                 scalars: Vec::new(),
                 terms: engine.resolve_terms(&descriptors),
@@ -729,7 +729,7 @@ fn a_reboot_after_a_merge_reads_back_the_watermark_the_process_served() {
                 UnallocatedRow {
                     external_id: Some(format!("late-{t}").into_bytes()),
                     view: "s0".to_string(),
-                    x: ((t * TIER_WIDTH) * 20) as f32,
+                    x: ((t * TIER_WIDTH) * 20) as f64,
                     y: 45.0,
                     scalars: Vec::new(),
                     terms: engine.resolve_terms(&descriptors),
@@ -851,7 +851,7 @@ fn flush_one_segment(engine: &Engine, tag: usize, rows: usize) -> Vec<EntityId> 
         batch.push(UnallocatedRow {
             external_id: Some(external_id.into_bytes()),
             view: "s0".to_string(),
-            x: ((t % 47) * 20) as f32,
+            x: ((t % 47) * 20) as f64,
             y: 5.0,
             scalars: Vec::new(),
             terms: engine.resolve_terms(&descriptors),
