+++ crates/tessera-store/examples/maintenance_memory.rs
@@ -97,8 +97,8 @@ fn write_inputs(root: &Path, rows: u64, segments: u64) -> Vec<MergeInput> {
                 .map(|e| FlushRow {
                     entity_id: EntityId::new(e),
                     external_id: Some(format!("ext-{e}").into_bytes()),
-                    x: (((e * 7 + s * 13) % 65_521) as f32) / 65_521.0,
-                    y: (((e * 31 + s * 17) % 65_519) as f32) / 65_519.0,
+                    x: (((e * 7 + s * 13) % 65_521) as f64) / 65_521.0,
+                    y: (((e * 31 + s * 17) % 65_519) as f64) / 65_519.0,
                     scalars: vec![],
                 })
                 .collect();
