+++ crates/tessera-store/tests/flush_segment.rs
@@ -21,7 +21,7 @@ fn unit_quantisation() -> Quantisation {
     }
 }
 
-fn row(entity: u64, external_id: Option<&[u8]>, x: f32, y: f32) -> FlushRow {
+fn row(entity: u64, external_id: Option<&[u8]>, x: f64, y: f64) -> FlushRow {
     FlushRow {
         entity_id: EntityId::new(entity),
         external_id: external_id.map(|id| id.to_vec()),
