+++ crates/tessera-store/tests/merge_execution.rs
@@ -33,8 +33,8 @@ fn segment(root: &Path, seg_id: &str, entity_lo: u64, count: u64, stride: u64) -
         .map(|e| FlushRow {
             entity_id: EntityId::new(e),
             external_id: Some(format!("ext-{e}").into_bytes()),
-            x: (((e * stride) % 97) as f32) / 97.0,
-            y: (((e * 53) % 89) as f32) / 89.0,
+            x: (((e * stride) % 97) as f64) / 97.0,
+            y: (((e * 53) % 89) as f64) / 89.0,
             scalars: vec![],
         })
         .collect();
